src-tauri/resources/ocr-worker/requirements.txt
rapidocr_onnxruntime
onnxruntime
Pillow
numpy
pyinstaller
